Fix score bug
The new score method returns the difference in score. The difference
was being directly compared to the old score, which resulted no circles
being saved after the first epoch. Treating the value as a difference
and adding it to the old score to get the new score fixes this issue.