Transmit circles between users
After a circle is generated by a user's frontend, it is sent through
as websocket to the backend server, where the circle is broadcasted to
every connected user. When a user receives a circle from the socket,
they draw it to the canvas.

The project has been restructured into a workspace, so that dependencies
between wasm and the backend server do not conflict.