Improve output
The output resolution was decreased to speed up the output,
but a varying mutation rate was added. This rate decreases over the
course of an epoch, allowing for finer adjustments at the end of an
epoch. The result is circles which score higher than before,
reducing the number of discarded circles overall.